refactor(attendance): AttendanceDetailModal sous 300 lignes (H7)
Éclatement complet (pattern G1) : la modale (604 l) devient un wrapper mince
(191 l, API parent inchangée) relayant overlay/transition/états vers trois
sous-composants présentationnels — AttendanceModalHeader (148), Table (235),
Footer (120). CSS déplacé verbatim par section. formatDuration non convergé
gardé local (#23) dans header+table (duplication assumée). Test de montage G7
préexistant conservé ; couverture sous-composants ajoutée (AttendanceModalParts).
